Fix blender26-meshio MQO exporter for Blender 2.72b
The blender26-meshio MQO exporter fails in Blender version 2.72b because
mesh.faces was removed from the API and replaced with mesh.polygons
about, I think, Blender 2.63.

MQO exporter now writes the "visible" property for objects in the MQO.

The absence of this property is no problem for Metasequoia but is needed
for my home made MQO readers.